split the api
