+++ wtypes/rdf.py
@@ -1,0 +1,5 @@
+"""RDF properties from wtypes
+
+Note
+----
+Requires rdflib"""
