+++ wtypes/widgets.py
@@ -1,0 +1,1 @@
+"""Widgets features for wtypes."""
